added exercises 4.13-4.22

+++ osa4/controllers/blogs.js
@@ -1,24 +1,33 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
-blogRouter.get('/', (request, response) => {
-	Blog
-		.find({})
-		.then(blogs => {
-			response.json(blogs)
-		})
+
+blogRouter.get('/', async (request, response) => {
+	const blogs = await Blog
+		.find({}).populate('user', { username: 1, name: 1})
+
+	response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
 	const body = request.body
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	await console.log(decodedToken)
+	const user = await request.user
 
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
 		url: body.url,
-		likes: body.likes || 0
+		likes: body.likes || 0,
+		user: user._id
 	})
 	const savedBlog = await blog.save()
+	user.blogs = user.blogs.concat(savedBlog._id)
+	await user.save()
+
 	response.status(201).json(savedBlog)
 })
 
@@ -31,19 +40,29 @@ blogRouter.get('/:id', async (request, response) => {
 	}
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+	const blog = await Blog.findById(request.params.id)
+	const user = await request.user
+	if (!(blog.user.toString() === user._id.toString())) {
+		return response.status(401).json({ error: 'token invalid' })
+	}
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
 
-blogRouter.put('/:id', async (request, response) => {
+blogRouter.put('/:id', middleware.userExtractor, async (request, response) => {
 	const body = request.body
-
+	const blogFinder = await Blog.findById(request.params.id)
+	const user = await request.user
+	if (!(user._id.toString() === blogFinder.user.toString())) {
+		return response.status(401).json({ error: 'token invalid' })
+	}
 	const blog = {
 		title: body.title,
 		author: body.author,
 		url: body.url,
-		likes: body.likes
+		likes: body.likes,
+		user: user._id
 	}
 
 	const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
